Rename files instead of exported objects
