+++ src/router.ts
@@ -4,11 +4,11 @@ import { globalErrorHandlerMiddleware } from "@middleware/globalErrorHandler.js"
 import { authenticationMiddleware } from "@middleware/authentication.js";
 import { rateLimiterMiddleware } from "@middleware/rateLimiter.js";
 import { checkServiceHealthOperation } from "@operations/checkServiceHealth.js";
-import { retrieveTemplateOperation } from "@operations/retrieveTemplate.js";
-import { retrieveNewSubmissionsOperation } from "@operations/retrieveNewSubmissions.js";
-import { retrieveSubmissionOperation } from "@operations/retrieveSubmission.js";
-import { confirmSubmissionOperation } from "@operations/confirmSubmission.js";
-import { reportSubmissionOperation } from "@operations/reportSubmission.js";
+import { retrieveTemplateOperationV1 } from "@operations/retrieveTemplate.v1.js";
+import { retrieveNewSubmissionsOperationV1 } from "@operations/retrieveNewSubmissions.v1.js";
+import { retrieveSubmissionOperationV1 } from "@operations/retrieveSubmission.v1.js";
+import { confirmSubmissionOperationV1 } from "@operations/confirmSubmission.v1.js";
+import { reportSubmissionOperationV1 } from "@operations/reportSubmission.v1.js";
 import { versionMiddleware } from "@middleware/version.js";
 
 import type {
@@ -23,25 +23,25 @@ export function buildRouter(): Router {
   const templateRoute = Router(INHERIT_PARAMS).get(
     "/",
     versionMiddleware(1),
-    operationHandler(retrieveTemplateOperation.v1),
+    operationHandler(retrieveTemplateOperationV1),
   );
 
   const newRoute = Router(INHERIT_PARAMS).get(
     "/",
     versionMiddleware(1),
-    operationHandler(retrieveNewSubmissionsOperation.v1),
+    operationHandler(retrieveNewSubmissionsOperationV1),
   );
 
   const confirmRoute = Router(INHERIT_PARAMS).put(
     "/:confirmationCode([a-f0-9]{8}-[a-f0-9]{4}-4[a-f0-9]{3}-[89ab][a-f0-9]{3}-[a-f0-9]{12})",
     versionMiddleware(1),
-    operationHandler(confirmSubmissionOperation.v1),
+    operationHandler(confirmSubmissionOperationV1),
   );
 
   const problemRoute = Router(INHERIT_PARAMS).post(
     "/",
     versionMiddleware(1),
-    operationHandler(reportSubmissionOperation.v1),
+    operationHandler(reportSubmissionOperationV1),
   );
 
   const submissionNameRoute = Router(INHERIT_PARAMS)
@@ -50,7 +50,7 @@ export function buildRouter(): Router {
     .get(
       "/",
       versionMiddleware(1),
-      operationHandler(retrieveSubmissionOperation.v1),
+      operationHandler(retrieveSubmissionOperationV1),
     );
 
   const submissionRoute = Router(INHERIT_PARAMS)
